// projects/bitsui-base/src/lib/types/fuse-navigation.ts
export interface FuseNavigationItem {
    id: string;
    title: string;
    type: "item" | "group" | "collapsable";
    translate?: string;
    icon?: string;
    hidden?: boolean;
    url?: string;
    classes?: string;
    exactMatch?: boolean;
    externalUrl?: boolean;
    openInNewTab?: boolean;
    isFeatureFound?: boolean;
    function?: any;
    badge?: {
        title?: string;
        translate?: string;
        bg?: string;
        fg?: string;
    };
    children?: FuseNavigationItem[];
}

export interface FuseNavigation extends FuseNavigationItem {
    children?: FuseNavigationItem[];
}